[Refactor:System] Organizing setup_sample_courses.py (#9753)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant

### What is the current behavior?
There is approximately 2250 lines in setup_sample_courses.py, which is
inconvenient to attach more/new functionality too and harder to
troubleshoot
### What is the new behavior?
The file has been split up in python packages and files that is more
manageable
