+++ .flake8
@@ -8,7 +8,14 @@ exclude=
     .setup/bin/reset_system.py,
     .setup/bin/reupload_generate_csv.py,
     .setup/bin/reupload_old_assignments.py,
-    .setup/bin/setup_sample_courses.py,
+    .setup/bin/sample_courses/models/__init__.py,
+    .setup/bin/sample_courses/models/component.py,
+    .setup/bin/sample_courses/models/gradeable.py,
+    .setup/bin/sample_courses/models/course/course_create_gradeables.py,
+    .setup/bin/sample_courses/models/course/course_create.py,
+    .setup/bin/sample_courses/models/course/course_data.py,
+    .setup/bin/sample_courses/models/course/course_generate_utils.py,
+    .setup/bin/sample_courses/models/course/course_utils.py,
     .setup/bin/track_git_version.py,
     .setup/CONFIGURE_SUBMITTY.py,
     .setup/update_database.py,
